feat: market snapshotting  (#38)
adds a package named `aggregator` that serves a purpose to snapshot the
market in fixed timeframes (4 hours, daily and weekly).
market has subcollections that store every events values in separate
docs. if we were to used these collection for graph purposes for example
we could end up in a situation where a market has an average 1000 loan
token supplies in an hour (figuratively speaking) which would result in
24 000 points in a graph that display the total supply in a day (let
alone week or month).

this is why market is being snapshot on a fixed time frame, so these
snapshots can be a representation of whatever timeframe (with these 3
precisions) without overloading the client side